requirements.txt
PyYAML==6.0.2
requests==2.32.3
prometheus-client==0.20.0
Jinja2==3.1.4

